+++ libs/textarea/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.1.0-next.1](https://github.com/spryker/ui-components/compare/@spryker/textarea@0.1.0-next.0...@spryker/textarea@0.1.0-next.1) (2020-03-12)
+
+
+### Bug Fixes
+
+* **textarea:** dependencies ([346961b](https://github.com/spryker/ui-components/commit/346961bf77906ff9133339263bb7d48ad90ddec2))
+
+
+
+
+
 # 0.1.0-next.0 (2020-02-27)